+++ artiq/coredevice/ad9912.py
@@ -182,3 +182,13 @@ class AD9912:
         """
         self.set_mu(self.frequency_to_ftw(frequency),
             self.turns_to_pow(phase))
+
+    @kernel
+    def cfg_sw(self, state):
+        """Set CPLD CFG RF switch state. The RF switch is controlled by the
+        logical or of the CPLD configuration shift register
+        RF switch bit and the SW TTL line (if used).
+
+        :param state: CPLD CFG RF switch bit
+        """
+        self.cpld.cfg_sw(self.chip_select - 4, state)