+++ artiq/test/coredevice/test_rtio.py
@@ -29,6 +29,23 @@ class RTIOCounter(EnvExperiment):
         self.set_dataset("dt", self.core.mu_to_seconds(t1 - t0))
 
 
+class InvalidCounter(Exception):
+    pass
+
+
+class WaitForRTIOCounter(EnvExperiment):
+    def build(self):
+        self.setattr_device("core")
+
+    @kernel
+    def run(self):
+        self.core.break_realtime()
+        target_mu = now_mu() + 10000
+        self.core.wait_until_mu(target_mu)
+        if self.core.get_rtio_counter_mu() < target_mu:
+            raise InvalidCounter
+
+
 class PulseNotReceived(Exception):
     pass
 
@@ -51,7 +68,7 @@ class RTT(EnvExperiment):
                 delay(1*us)
                 t0 = now_mu()
                 self.ttl_inout.pulse(1*us)
-        t1 = self.ttl_inout.timestamp_mu()
+        t1 = self.ttl_inout.timestamp_mu(now_mu())
         if t1 < 0:
             raise PulseNotReceived()
         self.set_dataset("rtt", self.core.mu_to_seconds(t1 - t0))
@@ -75,7 +92,7 @@ class Loopback(EnvExperiment):
                 delay(1*us)
                 t0 = now_mu()
                 self.loop_out.pulse(1*us)
-        t1 = self.loop_in.timestamp_mu()
+        t1 = self.loop_in.timestamp_mu(now_mu())
         if t1 < 0:
             raise PulseNotReceived()
         self.set_dataset("rtt", self.core.mu_to_seconds(t1 - t0))
@@ -98,7 +115,7 @@ class ClockGeneratorLoopback(EnvExperiment):
             with sequential:
                 delay(200*ns)
                 self.loop_clock_out.set(1*MHz)
-        self.set_dataset("count", self.loop_clock_in.count())
+        self.set_dataset("count", self.loop_clock_in.count(now_mu()))
 
 
 class PulseRate(EnvExperiment):
@@ -186,7 +203,7 @@ class LoopbackCount(EnvExperiment):
                 for i in range(self.npulses):
                     delay(25*ns)
                     self.loop_out.pulse(25*ns)
-        self.set_dataset("count", self.loop_in.count())
+        self.set_dataset("count", self.loop_in.count(now_mu()))
 
 
 class IncorrectLevel(Exception):
@@ -374,6 +391,9 @@ class CoredeviceTest(ExperimentCase):
         print(dt)
         self.assertGreater(dt, 50*ns)
         self.assertLess(dt, 1*us)
+
+    def test_wait_for_rtio_counter(self):
+        self.execute(WaitForRTIOCounter)
 
     def test_loopback(self):
         self.execute(Loopback)